requirements.txt
Adafruit-CharLCD==1.1.1
Adafruit-GPIO==1.0.4
Adafruit-PureIO==0.2.3
asn1crypto==0.24.0
asyncio==3.4.3
certifi==2018.8.24
cffi==1.13.1
chardet==3.0.4
cryptography==2.6.1
entrypoints==0.3
idna==2.6
keyring==17.1.1
keyrings.alt==3.1.1
numpy==1.16.2
PyAudio==0.2.11
pycparser==2.19
pycrypto==2.6.1
PyGObject==3.30.4
pymongo==3.4.0
python-apt==1.8.4
python-mpd2==1.0.0
pyxdg==0.25
requests==2.21.0
RPi.GPIO==0.7.0
RPLCD==1.2.2
SecretStorage==2.3.1
six==1.12.0
sounddevice==0.3.14
spidev==3.4
ssh-import-id==5.7
urllib3==1.24.1